Moved a function from ametice_bot.py to filename_parser.py and another to dic_operations.py
- Renamed download function in main.py module to main
- Moved get_cm_folder_path to filename_parser.py and get_classified_cm_id to dic_operations.py
