+++ entbot/main.py
@@ -9,7 +9,17 @@ from entbot.constants import Headers
 from entbot.tools.filename_parser import turn_cwd_to_execution_dir
 from entbot.tools.logging_config import display_message
 
-async def download(username, password):
+
+async def main(username: str, password: str):
+    """The main function to execute for downloading all ametice files
+    from the account associated to the given credentials.
+
+    Args:
+        - username (str): The username to sign in on the Aix Marseille website.
+        - password (str): The password to sign in on the Aix Marseille website.
+
+    Returns: None
+    """
     async with aiohttp.ClientSession(
         headers=Headers.LOGIN_HEADERS,
         connector=aiohttp.TCPConnector(force_close=True),
@@ -40,4 +50,4 @@ password = getpass(prompt="Mot de passe : ")
 print("")
 
 turn_cwd_to_execution_dir()
-asyncio.run(download(username, password))
+asyncio.run(main(username, password))
